db structure and adapter update

+++ app/src/main/java/com/ilih/weatherus/data/source/db/entity/DailyForecastEntity.kt
@@ -22,7 +22,7 @@ data class DailyForecastEntity(
     val weatherDesc: String,
     @ColumnInfo(name = "timestamp")
     val timestamp: Long,
-    val homeParent: Int
+    val homeParent: Long
 )
 
 fun DailyForecastEntity.toDto() =
